+++ R/useful-items.R
@@ -68,6 +68,7 @@ bs4Badge <- function(..., color, position = c("left", "right"),
 #' @param ... slot for \link{accordionItem}.
 #' @param id Unique accordion id.
 #' @param width The width of the accordion.
+#' @param .list To pass \link{accordionItem} within a list.
 #'
 #' @author David Granjon, \email{dgranjon@@ymail.com}
 #' @rdname accordion
@@ -111,18 +112,37 @@ bs4Badge <- function(..., color, position = c("left", "right"),
 #'           collapsed = FALSE,
 #'           "This is some text!"
 #'         )
-#'       )
+#'       ),
+#'       accordion(
+#'         id = "accordion_dynamic",
+#'         .list = lapply(
+#'           1:2,
+#'           function(i)
+#'             accordionItem(
+#'               title = paste('Accordion 1 Item', i),
+#'               status = "danger",
+#'               collapsed = ifelse (i == 1, TRUE, FALSE),
+#'               "This is some text!"
+#'             )
+#'          )
+#'        )
 #'     ),
 #'     title = "Accordion"
 #'   ),
-#'   server = function(input, output) { }
+#'   server = function(input, output) {
+#'    observe({
+#'      print(input$accordion1)
+#'      print(input$accordion2)
+#'      print(input$accordion_dynamic)
+#'    })
+#'   }
 #'  )
 #' }
 #'
 #' @export
-bs4Accordion <- function(..., id, width = 12) {
+bs4Accordion <- function(..., id, width = 12, .list = NULL) {
   
-  items <- list(...)
+  items <- c(list(...), .list)
   
   # patch that enables a proper accordion behavior
   # we add the data-parent non standard attribute to each